+++ backend/src/main/java/com/rikai/backend/dto/request/weeklyreport/WeeklyReportCreateDTO.java
@@ -9,28 +9,28 @@ import java.time.LocalDate;
 
 @Data
 public class WeeklyReportCreateDTO {
-    
+
     @NotNull(message = "INTERN_ID_REQUIRED")
-    private Integer internId;
-    
+    private Long internId;
+
     @NotNull(message = "WEEK_START_DATE_REQUIRED")
     private LocalDate weekStartDate;
-    
+
     private String tasksAssigned;
-    
+
     private String tasksCompleted;
-    
+
     private String outputQuality;
-    
+
     @Min(value = 1, message = "INVALID_SCORE")
     @Max(value = 10, message = "INVALID_SCORE")
     private Byte proactivityScore;
-    
+
     @Min(value = 1, message = "INVALID_SCORE")
     @Max(value = 10, message = "INVALID_SCORE")
     private Byte progressScore;
-    
+
     private String issuesRisks;
-    
+
     private String mentorOverallComment;
 }
